+++ Shooter.cpp
@@ -1,6 +1,6 @@
 #include "Shooter.h"
 
-Shooter::Shooter(int x, int y, float reloadTime, int bulletSpeed, string bulletSpritePath, double textureX, double textureY) :pos(x, y), reloadTime(reloadTime), bulletSpeed(bulletSpeed), bulletSpritePath(bulletSpritePath),textureX(textureX),textureY(textureY) {
+Shooter::Shooter(int x, int y, double reloadTime, int bulletSpeed, string bulletSpritePath, double textureX, double textureY) :pos(x, y), reloadTime(reloadTime), bulletSpeed(bulletSpeed), bulletSpritePath(bulletSpritePath),textureX(textureX),textureY(textureY) {
 	bullet = nullptr;
 }
